feat(desktop): Add context settings to toggle docker/git credential forwarding behaviour for default context from desktop app
