+++ desktop/src/types.ts
@@ -206,7 +206,11 @@ export type TImportWorkspaceConfig = Readonly<{
 //#region Context
 export type TContextOptions = Record<TContextOptionName, TContextOption>
 // See pkg/config/context.go
-export type TContextOptionName = "AGENT_URL" | "TELEMETRY"
+export type TContextOptionName =
+  | "AGENT_URL"
+  | "TELEMETRY"
+  | "SSH_INJECT_DOCKER_CREDENTIALS"
+  | "SSH_INJECT_GIT_CREDENTIALS"
 export type TContextOption = Readonly<{
   name: TContextOptionName
   description: string | null | undefined
