+++ desktop/src/views/Settings/useContextOptions.tsx
@@ -1,21 +1,10 @@
-import { CloseIcon } from "@chakra-ui/icons"
-import {
-  Code,
-  IconButton,
-  Input,
-  InputGroup,
-  InputLeftAddon,
-  InputRightElement,
-  Link,
-  Switch,
-} from "@chakra-ui/react"
+import { Code, Link, Switch } from "@chakra-ui/react"
 import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query"
-import { FocusEvent, KeyboardEvent, ReactNode, useCallback, useMemo, useRef, useState } from "react"
+import { useCallback, useMemo } from "react"
 import { client } from "../../client"
-import { Command } from "../../client/command"
-import { TSettings, useChangeSettings } from "../../contexts"
 import { QueryKeys } from "../../queryKeys"
 import { TContextOptionName } from "../../types"
+import { ClearableInput } from "./ClearableInput"
 
 const DEFAULT_DEVPOD_AGENT_URL = "https://github.com/loft-sh/devpod/releases/latest/download/"
 
@@ -41,33 +30,6 @@ export function useContextOptions() {
     }),
     [options, updateOption]
   )
-}
-export function useCLIFlagsOption() {
-  const { settings, set } = useChangeSettings()
-
-  const handleChanged = useCallback(
-    (newValue: string) => {
-      const value = newValue.trim()
-      set("additionalCliFlags", value)
-      client.setSetting("additionalCliFlags", value)
-    },
-    [set]
-  )
-
-  const input = useMemo(
-    () => (
-      <ClearableInput
-        placeholder="Additional CLI Flags"
-        defaultValue={settings.additionalCliFlags}
-        onChange={handleChanged}
-      />
-    ),
-    [handleChanged, settings.additionalCliFlags]
-  )
-
-  const helpText = useMemo(() => <>Set additional CLI Flags to use.</>, [])
-
-  return { input, helpText }
 }
 
 export function useAgentURLOption() {
@@ -104,6 +66,7 @@ export function useAgentURLOption() {
 
   return { input, helpText }
 }
+
 export function useTelemetryOption() {
   const { options, updateOption } = useContextOptions()
 
@@ -134,203 +97,54 @@ export function useTelemetryOption() {
   return { input, helpText }
 }
 
-export function useExtraEnvVarsOption() {
-  const { settings, set } = useChangeSettings()
-
-  const handleChanged = useCallback(
-    (newValue: string) => {
-      const value = newValue.trim()
-      set("additionalEnvVars", value)
-      Command.ADDITIONAL_ENV_VARS = value
-    },
-    [set]
-  )
+export function useDockerCredentialsForwardingOption() {
+  const { options, updateOption } = useContextOptions()
 
   const input = useMemo(
     () => (
-      <ClearableInput
-        placeholder="Additional Environment Variables"
-        defaultValue={settings.additionalEnvVars}
-        onChange={handleChanged}
+      <Switch
+        isChecked={options?.SSH_INJECT_DOCKER_CREDENTIALS.value === "true"}
+        onChange={(e) =>
+          updateOption({
+            option: "SSH_INJECT_DOCKER_CREDENTIALS",
+            value: e.target.checked.toString(),
+          })
+        }
       />
     ),
-    [handleChanged, settings.additionalEnvVars]
+    [options?.SSH_INJECT_DOCKER_CREDENTIALS.value, updateOption]
   )
 
   const helpText = useMemo(
-    () => (
-      <>
-        Set additional environment variables DevPod passes to all commands. Accepts a comma
-        separated list, e.g. FOO=bar,BAZ=false
-      </>
-    ),
+    () => <>Enable to forward your local docker credentials to workspaces</>,
     []
   )
 
   return { input, helpText }
 }
 
-export function useDotfilesOption() {
-  const { settings, set } = useChangeSettings()
-  const updateOption = useCallback(
-    (value: string) => {
-      set("dotfilesUrl", value)
-      client.setSetting("dotfilesUrl", value)
-    },
-    [set]
-  )
+export function useGitCredentialsForwardingOption() {
+  const { options, updateOption } = useContextOptions()
 
   const input = useMemo(
     () => (
-      <ClearableInput
-        placeholder="Dotfiles repo URL"
-        defaultValue={settings.dotfilesUrl}
-        onChange={updateOption}
+      <Switch
+        isChecked={options?.SSH_INJECT_GIT_CREDENTIALS.value === "true"}
+        onChange={(e) =>
+          updateOption({
+            option: "SSH_INJECT_GIT_CREDENTIALS",
+            value: e.target.checked.toString(),
+          })
+        }
       />
     ),
-    [settings.dotfilesUrl, updateOption]
+    [options?.SSH_INJECT_GIT_CREDENTIALS.value, updateOption]
   )
-
-  return { input }
-}
-
-export function useSSHKeySignatureOption() {
-  const { settings, set } = useChangeSettings()
-  const updateOption = useCallback(
-    (value: string) => {
-      set("sshKeyPath", value)
-      client.setSetting("sshKeyPath", value)
-    },
-    [set]
-  )
-
-  const input = useMemo(
-    () => (
-      <ClearableInput
-        placeholder="SSH key path"
-        defaultValue={settings.sshKeyPath}
-        onChange={updateOption}
-      />
-    ),
-    [settings.sshKeyPath, updateOption]
-  )
-
-  return { input }
-}
-
-export function useProxyOptions() {
-  const { settings, set } = useChangeSettings()
-
-  const handleChanged = useCallback(
-    (s: keyof Pick<TSettings, "httpProxyUrl" | "httpsProxyUrl" | "noProxy">) =>
-      (newValue: string) => {
-        set(s, newValue)
-        client.setSetting(s, newValue)
-      },
-    [set]
-  )
-
-  const input = useMemo(() => {
-    return (
-      <>
-        <ClearableInput
-          defaultValue={settings.httpProxyUrl}
-          onChange={handleChanged("httpProxyUrl")}
-          placeholder="Set HTTP_PROXY"
-          label="HTTP"
-        />
-        <ClearableInput
-          defaultValue={settings.httpsProxyUrl}
-          onChange={handleChanged("httpsProxyUrl")}
-          placeholder="Set HTTPS_PROXY"
-          label="HTTPS"
-        />
-        <ClearableInput
-          defaultValue={settings.noProxy}
-          onChange={handleChanged("noProxy")}
-          placeholder="Set NO_PROXY"
-          label="NO_PROXY"
-        />
-      </>
-    )
-  }, [handleChanged, settings.httpProxyUrl, settings.httpsProxyUrl, settings.noProxy])
 
   const helpText = useMemo(
-    () => (
-      <>
-        Set HTTP(S) proxy configuration. These settings will only be used by DevPod itself and not
-        be available within your workspace.
-      </>
-    ),
+    () => <>Enable to forward your local HTTPS based git credentials to workspaces</>,
     []
   )
 
   return { input, helpText }
-}
-
-type TClearableInputProps = Readonly<{
-  defaultValue: string
-  placeholder: string
-  label?: ReactNode
-  onChange: (newValue: string) => void
-}>
-function ClearableInput({ defaultValue, placeholder, label, onChange }: TClearableInputProps) {
-  const [hasFocus, setHasFocus] = useState(false)
-  const inputRef = useRef<HTMLInputElement>(null)
-
-  const handleBlur = useCallback(
-    (e: FocusEvent<HTMLInputElement>) => {
-      const value = e.target.value.trim()
-      onChange(value)
-      setHasFocus(false)
-    },
-    [onChange]
-  )
-
-  const handleKeyUp = useCallback((e: KeyboardEvent<HTMLInputElement>) => {
-    if (e.key !== "Enter") return
-
-    e.currentTarget.blur()
-  }, [])
-
-  const handleFocus = useCallback(() => {
-    setHasFocus(true)
-  }, [])
-
-  const handleClearClicked = useCallback(() => {
-    const el = inputRef.current
-    if (!el) return
-
-    el.value = ""
-  }, [])
-
-  return (
-    <InputGroup maxWidth="96">
-      {label && <InputLeftAddon>{label}</InputLeftAddon>}
-      <Input
-        ref={inputRef}
-        spellCheck={false}
-        placeholder={placeholder}
-        defaultValue={defaultValue}
-        onBlur={handleBlur}
-        onKeyUp={handleKeyUp}
-        onFocus={handleFocus}
-      />
-      <InputRightElement>
-        <IconButton
-          visibility={hasFocus ? "visible" : "hidden"}
-          size="xs"
-          borderRadius="full"
-          icon={<CloseIcon />}
-          aria-label="clear"
-          onMouseDown={(e) => {
-            // needed to prevent losing focus from input
-            e.stopPropagation()
-            e.preventDefault()
-          }}
-          onClick={handleClearClicked}
-        />
-      </InputRightElement>
-    </InputGroup>
-  )
 }
